+++ packages/categorize/configure/src/design/index.jsx
@@ -1,9 +1,14 @@
+import { getPluginProps } from './utils';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import { FeedbackConfig, InputContainer, layout, settings } from '@pie-lib/pie-toolbox/config-ui';
-import { countInAnswer, ensureNoExtraChoicesInAnswer, ensureNoExtraChoicesInAlternate } from '@pie-lib/pie-toolbox/categorize';
-import {EditableHtml} from '@pie-lib/pie-toolbox/editable-html';
+import {
+  countInAnswer,
+  ensureNoExtraChoicesInAnswer,
+  ensureNoExtraChoicesInAlternate,
+} from '@pie-lib/pie-toolbox/categorize';
+import { EditableHtml } from '@pie-lib/pie-toolbox/editable-html';
 import { uid, withDragContext } from '@pie-lib/pie-toolbox/drag';
 
 import Categories from './categories';
@@ -81,7 +86,9 @@ export class Design extends React.Component {
     }));
 
     // ensure that maxChoicesPerCategory is reset if author switch back the corresponding switch (allowMaxChoicesPerCategory)
-    updatedModel.maxChoicesPerCategory = updatedModel.allowMaxChoicesPerCategory ? updatedModel.maxChoicesPerCategory : 0;
+    updatedModel.maxChoicesPerCategory = updatedModel.allowMaxChoicesPerCategory
+      ? updatedModel.maxChoicesPerCategory
+      : 0;
 
     onChange(updatedModel);
   };
@@ -168,12 +175,11 @@ export class Design extends React.Component {
     this.updateModel({ maxChoicesPerCategory: maxChoices });
   };
 
-
   render() {
     const { classes, configuration, imageSupport, model, uploadSoundSupport, onConfigurationChanged } = this.props;
     const {
       allowMultiplePlacements = {},
-      allowAlternate = {},
+      baseInputConfiguration = {},
       categoriesPerRow = {},
       choicesPosition = {},
       contentDimensions = {},
@@ -199,6 +205,7 @@ export class Design extends React.Component {
     const {
       allowAlternateEnabled,
       allowMaxChoicesPerCategory,
+      errors,
       feedbackEnabled,
       maxChoicesPerCategory,
       promptEnabled,
@@ -207,6 +214,11 @@ export class Design extends React.Component {
       teacherInstructionsEnabled,
       toolbarEditorPosition,
     } = model || {};
+    const {
+      prompt: promptError,
+      rationale: rationaleError,
+      teacherInstructions: teacherInstructionsError,
+    } = errors || {};
 
     const toolbarOpts = {
       position: toolbarEditorPosition === 'top' ? 'top' : 'bottom',
@@ -226,7 +238,7 @@ export class Design extends React.Component {
     const choices = model.choices.map((c) => {
       c.correctResponseCount = this.countChoiceInCorrectResponse(c);
       // ensure categoryCount is set even though updatedModel hasn't been called
-      c.categoryCount =  this.checkAllowMultiplePlacements(model.allowMultiplePlacementsEnabled, c);
+      c.categoryCount = this.checkAllowMultiplePlacements(model.allowMultiplePlacementsEnabled, c);
       return c;
     });
 
@@ -277,7 +289,10 @@ export class Design extends React.Component {
     };
 
     const isOpened = this.isAlertModalOpened();
-    const alertMaxChoicesMsg = translator.t('translation:categorize:maxChoicesPerCategoryRestriction', {lng: model.language, maxChoicesPerCategory });
+    const alertMaxChoicesMsg = translator.t('translation:categorize:maxChoicesPerCategoryRestriction', {
+      lng: model.language,
+      maxChoicesPerCategory,
+    });
 
     return (
       <IdProvider value={this.uid}>
@@ -294,12 +309,14 @@ export class Design extends React.Component {
                 Settings: panelSettings,
                 Properties: panelProperties,
               }}
-              modal={<AlertDialog
+              modal={
+                <AlertDialog
                   title={'Warning'}
                   text={alertMaxChoicesMsg}
                   open={isOpened}
                   onClose={this.onAlertModalCancel}
-              />}
+                />
+              }
             />
           }
         >
@@ -310,8 +327,10 @@ export class Design extends React.Component {
                 markup={model.teacherInstructions || ''}
                 onChange={this.changeTeacherInstructions}
                 imageSupport={imageSupport}
+                error={teacherInstructionsError}
                 nonEmpty={false}
                 toolbarOpts={toolbarOpts}
+                pluginProps={getPluginProps(teacherInstructions?.inputConfiguration, baseInputConfiguration)}
                 spellCheck={spellCheckEnabled}
                 maxImageWidth={(maxImageWidth && maxImageWidth.teacherInstructions) || defaultImageMaxWidth}
                 maxImageHeight={(maxImageHeight && maxImageHeight.teacherInstructions) || defaultImageMaxHeight}
@@ -319,6 +338,7 @@ export class Design extends React.Component {
                 languageCharactersProps={[{ language: 'spanish' }, { language: 'special' }]}
                 mathMlOptions={mathMlOptions}
               />
+              {teacherInstructionsError && <div className={classes.errorText}>{teacherInstructionsError}</div>}
             </InputContainer>
           )}
 
@@ -329,9 +349,11 @@ export class Design extends React.Component {
                 markup={model.prompt || ''}
                 onChange={this.onPromptChanged}
                 imageSupport={imageSupport}
+                error={promptError}
                 nonEmpty={false}
                 disableUnderline
                 toolbarOpts={toolbarOpts}
+                pluginProps={getPluginProps(prompt?.inputConfiguration, baseInputConfiguration)}
                 spellCheck={spellCheckEnabled}
                 maxImageWidth={maxImageWidth && maxImageWidth.prompt}
                 maxImageHeight={maxImageHeight && maxImageHeight.prompt}
@@ -339,6 +361,7 @@ export class Design extends React.Component {
                 languageCharactersProps={[{ language: 'spanish' }, { language: 'special' }]}
                 mathMlOptions={mathMlOptions}
               />
+              {promptError && <div className={classes.errorText}>{promptError}</div>}
             </InputContainer>
           )}
 
@@ -412,8 +435,10 @@ export class Design extends React.Component {
                 markup={model.rationale || ''}
                 onChange={this.changeRationale}
                 imageSupport={imageSupport}
+                error={rationaleError}
                 nonEmpty={false}
                 toolbarOpts={toolbarOpts}
+                pluginProps={getPluginProps(prompt?.inputConfiguration, baseInputConfiguration)}
                 spellCheck={spellCheckEnabled}
                 maxImageWidth={(maxImageWidth && maxImageWidth.rationale) || defaultImageMaxWidth}
                 maxImageHeight={(maxImageHeight && maxImageHeight.rationale) || defaultImageMaxHeight}
@@ -421,6 +446,7 @@ export class Design extends React.Component {
                 languageCharactersProps={[{ language: 'spanish' }, { language: 'special' }]}
                 mathMlOptions={mathMlOptions}
               />
+              {rationaleError && <div className={classes.errorText}>{rationaleError}</div>}
             </InputContainer>
           )}
 
@@ -448,6 +474,11 @@ const styles = (theme) => ({
   },
   title: {
     marginBottom: theme.spacing.unit * 4,
+  },
+  errorText: {
+    fontSize: theme.typography.fontSize - 2,
+    color: theme.palette.error.main,
+    paddingTop: theme.spacing.unit,
   },
 });