+++ packages/categorize/configure/src/design/input-header.jsx
@@ -1,13 +1,15 @@
+import { getPluginProps } from './utils';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import classNames from 'classnames';
-import {EditableHtml} from '@pie-lib/pie-toolbox/editable-html';
+import { EditableHtml } from '@pie-lib/pie-toolbox/editable-html';
 
 export class InputHeader extends React.Component {
   static propTypes = {
     classes: PropTypes.object.isRequired,
     className: PropTypes.string,
+    configuration: PropTypes.object.isRequired,
     deleteFocusedEl: PropTypes.func,
     disabled: PropTypes.bool,
     focusedEl: PropTypes.number,
@@ -46,6 +48,7 @@ export class InputHeader extends React.Component {
   render() {
     const {
       onChange,
+      configuration,
       label,
       classes,
       className,
@@ -58,12 +61,10 @@ export class InputHeader extends React.Component {
       maxImageWidth,
       maxImageHeight,
       uploadSoundSupport,
-      mathMlOptions = {}
+      mathMlOptions = {},
     } = this.props;
-    const choicePlugins = {
-      audio: { disabled: true },
-      video: { disabled: true },
-    };
+
+    const { headers, baseInputConfiguration } = configuration;
 
     return (
       <div className={classNames(classes.inputHeader, className)}>
@@ -76,17 +77,14 @@ export class InputHeader extends React.Component {
           markup={label}
           onChange={onChange}
           className={classes.editor}
-          pluginProps={choicePlugins}
+          pluginProps={getPluginProps(headers?.inputConfiguration, baseInputConfiguration)}
           toolbarOpts={toolbarOpts}
           spellCheck={spellCheck}
           error={error}
           maxImageWidth={maxImageWidth}
           maxImageHeight={maxImageHeight}
           uploadSoundSupport={uploadSoundSupport}
-          languageCharactersProps={[
-            { language: 'spanish' },
-            { language: 'special' },
-          ]}
+          languageCharactersProps={[{ language: 'spanish' }, { language: 'special' }]}
           onDone={() => {
             deleteFocusedEl && deleteFocusedEl();
           }}