+++ packages/categorize/configure/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.8.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize-configure@7.8.0...@pie-element/categorize-configure@7.8.1) (2024-03-13)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-3549 ([b5d7d13](https://github.com/pie-framework/pie-elements/commit/b5d7d1352169cc67eaedcd56898a5806b0a8812a))
+
+
+
+
+
+# [7.8.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize-configure@7.7.0...@pie-element/categorize-configure@7.8.0) (2024-03-12)
+
+
+### Features
+
+* Add configuration for custom buttons PD-3514 ([78c4add](https://github.com/pie-framework/pie-elements/commit/78c4add8bf3796577e7863814c3f9e93ba607231))
+
+
+
+
+
+# [7.7.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize-configure@7.6.1...@pie-element/categorize-configure@7.7.0) (2024-03-12)
+
+
+### Features
+
+* **categorize:** add required fields functionality PD-3546 ([e370f14](https://github.com/pie-framework/pie-elements/commit/e370f14ff25afff8154b1b08a4ad4626aadcc547))
+
+
+
+
+
+## [7.6.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize-configure@7.6.0...@pie-element/categorize-configure@7.6.1) (2024-03-12)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-3509, PD-3465 ([b1745aa](https://github.com/pie-framework/pie-elements/commit/b1745aa981722b5561aab1f4495ac81ac7bc4155))
+
+
+
+
+
 # [7.6.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize-configure@7.5.0...@pie-element/categorize-configure@7.6.0) (2024-03-04)