+++ packages/categorize/controller/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.5.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize-controller@6.5.0...@pie-element/categorize-controller@6.5.1) (2024-03-13)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-3549 ([b5d7d13](https://github.com/pie-framework/pie-elements/commit/b5d7d1352169cc67eaedcd56898a5806b0a8812a))
+
+
+
+
+
+# [6.5.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize-controller@6.4.1...@pie-element/categorize-controller@6.5.0) (2024-03-12)
+
+
+### Features
+
+* **categorize:** add required fields functionality PD-3546 ([e370f14](https://github.com/pie-framework/pie-elements/commit/e370f14ff25afff8154b1b08a4ad4626aadcc547))
+
+
+
+
+
+## [6.4.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize-controller@6.4.0...@pie-element/categorize-controller@6.4.1) (2024-03-12)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-3509, PD-3465 ([b1745aa](https://github.com/pie-framework/pie-elements/commit/b1745aa981722b5561aab1f4495ac81ac7bc4155))
+
+
+
+
+
 # [6.4.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize-controller@6.3.2...@pie-element/categorize-controller@6.4.0) (2024-03-04)