+++ packages/charting/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.6.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/charting@7.6.0...@pie-element/charting@7.6.1) (2024-03-13)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-3549 ([b5d7d13](https://github.com/pie-framework/pie-elements/commit/b5d7d1352169cc67eaedcd56898a5806b0a8812a))
+
+
+
+
+
+# [7.6.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/charting@7.5.1...@pie-element/charting@7.6.0) (2024-03-12)
+
+
+### Features
+
+* Add configuration for custom buttons PD-3515 ([645574c](https://github.com/pie-framework/pie-elements/commit/645574c42638d85253afa0aefe5d09fd8fae0207))
+
+
+
+
+
+## [7.5.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/charting@7.5.0...@pie-element/charting@7.5.1) (2024-03-12)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-3509, PD-3465 ([b1745aa](https://github.com/pie-framework/pie-elements/commit/b1745aa981722b5561aab1f4495ac81ac7bc4155))
+
+
+
+
+
 # [7.5.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/charting@7.4.2...@pie-element/charting@7.5.0) (2024-03-04)