+++ packages/image-cloze-association/configure/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.5.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/image-cloze-association-configure@5.5.0...@pie-element/image-cloze-association-configure@5.5.1) (2024-03-13)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-3549 ([b5d7d13](https://github.com/pie-framework/pie-elements/commit/b5d7d1352169cc67eaedcd56898a5806b0a8812a))
+
+
+
+
+
+# [5.5.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/image-cloze-association-configure@5.4.1...@pie-element/image-cloze-association-configure@5.5.0) (2024-03-12)
+
+
+### Features
+
+* Add configuration for custom buttons PD-3524 ([94d3fc9](https://github.com/pie-framework/pie-elements/commit/94d3fc9aa56ddaa41b3dbaa8bc6c2a01514b9202))
+
+
+
+
+
+## [5.4.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/image-cloze-association-configure@5.4.0...@pie-element/image-cloze-association-configure@5.4.1) (2024-03-12)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-3509, PD-3465 ([b1745aa](https://github.com/pie-framework/pie-elements/commit/b1745aa981722b5561aab1f4495ac81ac7bc4155))
+
+
+
+
+
 # [5.4.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/image-cloze-association-configure@5.3.2...@pie-element/image-cloze-association-configure@5.4.0) (2024-03-04)