+++ packages/match/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.2.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/match@7.2.1...@pie-element/match@7.2.2) (2024-03-14)
+
+
+### Bug Fixes
+
+* **match:** check for pie-author when configuration is set PD-3558 ([aaf93fe](https://github.com/pie-framework/pie-elements/commit/aaf93fe456b1ac98b09890561b478d18d4eaf6f0))
+
+
+
+
+
+## [7.2.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/match@7.2.0...@pie-element/match@7.2.1) (2024-03-13)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-3549 ([b5d7d13](https://github.com/pie-framework/pie-elements/commit/b5d7d1352169cc67eaedcd56898a5806b0a8812a))
+
+
+
+
+
+# [7.2.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/match@7.1.2...@pie-element/match@7.2.0) (2024-03-12)
+
+
+### Features
+
+* **match:** add required fields functionality PD-3548 ([469f52f](https://github.com/pie-framework/pie-elements/commit/469f52f5ac795991ce115b11c6eee19eaedbf851))
+
+
+
+
+
+## [7.1.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/match@7.1.1...@pie-element/match@7.1.2) (2024-03-12)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-3509, PD-3465 ([b1745aa](https://github.com/pie-framework/pie-elements/commit/b1745aa981722b5561aab1f4495ac81ac7bc4155))
+
+
+
+
+
 ## [7.1.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/match@7.1.0...@pie-element/match@7.1.1) (2024-03-07)