+++ packages/match/controller/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.1.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-controller@6.1.0...@pie-element/match-controller@6.1.1) (2024-03-13)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-3549 ([b5d7d13](https://github.com/pie-framework/pie-elements/commit/b5d7d1352169cc67eaedcd56898a5806b0a8812a))
+
+
+
+
+
+# [6.1.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-controller@6.0.2...@pie-element/match-controller@6.1.0) (2024-03-12)
+
+
+### Features
+
+* **match:** add required fields functionality PD-3548 ([469f52f](https://github.com/pie-framework/pie-elements/commit/469f52f5ac795991ce115b11c6eee19eaedbf851))
+
+
+
+
+
+## [6.0.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-controller@6.0.1...@pie-element/match-controller@6.0.2) (2024-03-12)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-3509, PD-3465 ([b1745aa](https://github.com/pie-framework/pie-elements/commit/b1745aa981722b5561aab1f4495ac81ac7bc4155))
+
+
+
+
+
 ## [6.0.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-controller@6.0.0...@pie-element/match-controller@6.0.1) (2024-03-07)